Update 01. Trapeziod Area.py

+++ PB - More Exercises Python/1.0 Simple Operations and Calculations - More Exercises/01. Trapeziod Area.py
@@ -3,9 +3,9 @@ print(f'{(b1 + b2) * h / 2:.2f}')
 
 #----------------------------(2)----------------------------
 
-b1 = int(input())
-b2 = int(input())
-h = int(input())
+b1 = float(input())
+b2 = float(input())
+h = float(input())
 area = (b1 + b2) * h / 2
 print(f'{area:.2f}')
